Wait for rook-ceph to be applied in factory backup stage
If rook-ceph is configured in a standalone system, during a factory
install the backup script fails when it tries to create the ceph
backup, since the rook-ceph application is not yet applied.

This commit updates the 10-system-backup script to wait for the
rook-ceph application to reach the "applied" state before running the
backup.

When rook-ceph is configured, this check adds between 7 to 11 minutes
to the factory-install process, depending on how long rook-ceph takes
to reach the applied state.

Test Plan:
PASS: Run a factory-install in a standalone host with rook-ceph
configured. Verify that the setup backup waits until the app is applied
and the factory-install completes successfully.
PASS: Run a factory-install in a standalone host without rook-ceph
configured. Verify that the backup stage does not wait for the app.

Closes-bug: 2125236

Change-Id: I67b18426b1acc3f3389f2aae71a93b7bb0a66175
Signed-off-by: Enzo Candotti <Enzo.Candotti@windriver.com>