Add new config key to get the restful certificate
The ceph mgr restful module certificate is set with the key
'mgr/restful/<mgr_name>/crt'. However, when rook-ceph is
configured, a race condition can occur between the creation
of the mgr and the execution of the 'ceph-mgr-provision' job,
causing restful to not running.

To avoid this, instead of using a specific mgr key, the global
key 'mgr/restful/crt' will be used, so there is no risk of
having the mgr without a certificate.

Therefore, it was necessary to make the adjustment in cephclient
to check if the key 'mgr/restful/crt' exists. If it does, the
certificate saved in it will be obtained. Otherwise, the same
key as before, which is 'mgr/restful/<mgr_active>/crt', will be
used.

Test Plan:
 - PASS: AIO-SX with rook-ceph backend
 - PASS: AIO-DX with rook-ceph backend
 - PASS: AIO-SX with ceph backend
 - PASS: AIO-DX with ceph backend

Partial-Bug: 2125605

Change-Id: I5a5ec4a77a359a1f6d72d2e714127f3b93e00cb2
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>