Update subcloud time as part of enroll init
After a factory install, the system can be powered off for extended
periods of time, making the subcloud enrollment to fail. To fix this
issue, the time on the servers will need to be updated during the
subcloud enrollment.

This commit updates the enroll-init-reconfigure script to receive
a new parameter from the system controller, that will be used to
update the hardware and the system clock. Also, the certificate check
flow was updated to ensure that the certificates renewed by
cert-manager are checked and updated if needed.

Test Plan:
1. PASS: Run subcloud enrollment without the enroll-timestamp
   parameter. Verify that the process is not modified and it finishes
   correctly.
2. PASS: Run subcloud enrollment in a subcloud staged for more than
   90 days, sending the enroll-timestamp parameter. Verify the system
   time is updated correctly, cert-manager renews the certificates,
   and the enrollment completes successfully.
3. PASS: Same test as 1 but with the subcloud staged for more than 365
   days.
4. PASS: Run subcloud enrollment with the enroll-timestamp parameter
   in a subcloud with correct datetime. Verify that the time is not
   updated since the gap is less than 20 seconds.

Partial-Bug: https://bugs.launchpad.net/starlingx/+bug/2126942

Change-Id: I75241290d68f230c2f30a4c8aaef45d5408c77a2
Signed-off-by: Enzo Candotti <Enzo.Candotti@windriver.com>