Delete system resource post factory install
The system resource may not be reconciled post factory-install in a
multi-node system as some of the configurations can only be configured
after installed the controller-1, e.g. controllerfs. Leaving the
system resource unreconciled creates problem that some of the new
configurations during enrollment may be overwritten back to the status
at the end of the factory-install.

This commit adds tasks to delete the system resource regardless its
status as a new system resource will be applied during subcloud
enrollment for further configuration.

This commit also corrects the namespace of the CRs getting from the
actual configuration file instead of the default ns, removes
restriction to have the CRs applied to the default namespace only.

Test plan:
1. Passed - factory install AIO-SX with "deployment" namespace,
verified the factory install successful, verified enrollment
successful.

2. Passed - factory install AIO-SX with "customized" namespace,
verified the factory install successful, verified enrollment
successful with all the CRs defined in the "customized" namespace.

3. Passed - factory install AIO-DX with private registry overrides
defined in the localhost.yml, and docker-distribution resized to 50G
in deployment-config.yaml (which can only be resized after both
controllers are available, causing the system CR cannot be reconciled).
Verified factory install successful with private registry in service
parameters, verified subcloud enrollment successful, verified
registries are updated to registry.central.

Closes-bug: 2128351

Change-Id: Ic5168a655729eb39a88923164360492c308e4191
Signed-off-by: Jiang, Yuxing <Yuxing.Jiang@windriver.com>