Update password masking rules in collect logs
The fixes from [1] and [2] were made to a file that is not deployed to
the system when installing the .deb package.

This change copies the two changes to the script in `debian-scripts`
from the `scripts` folder.

[1] https://review.opendev.org/c/starlingx/utilities/+/926550
[2] https://review.opendev.org/c/starlingx/utilities/+/949801

Test Plan:
PASS: build-pkgs -c -p collector
PASS: Install the collector and verify the script is updated in the
      system path
PASS: Run collect tool and check for docker login masked passwords

Closes-Bug: 2110929

Change-Id: Id5a509bf023866f80b0f6d838999f4add6fd78b6
Signed-off-by: Rodrigo Tavares <Rodrigo.dosSantosTavares@windriver.com>
(cherry picked from commit 46ee57e52d2f3cc9320b9b961fe1544b1fcf5075)