Collect ipset current state in the logs
With the delivery of [1], it is necessary to collect the ipset state
on the system. This change implements that functionality.

Additionally, it collects system netfilter information because
iptables does not provide detailed information about Kubernetes
service rules (specifically, it does not show the Layer 4 port
associated with the service). Since netfilter is the underlying
implementation, it provides this level of detail.

Another missing piece of information is the list of listening sockets
for TCP, UDP, and SCTP. The collection of sm.info currently provides
netstat output only for the main network namespace on controller
nodes. This change uses ss to collect socket information on all nodes
and pods that have the worker subfunction.

[1] https://review.opendev.org/c/starlingx/config/+/967532

Test Plan
=========
[PASS] collect a node's logs and verify new information in the file
       networking.info

Story: 2011311
Task: 53262

Depends-On: https://review.opendev.org/c/starlingx/tools/+/967539
Change-Id: I017d2d4c04f157223f997403cee3f8bd17370162
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>