Delay fm-api start until DNS is available
The fm-api must start as early as possible so it can process alarms from
other services. During a fresh install the fm-api host is a valid IP
address, but after unlock it receives a host FQDN (for example:
controller-0.internal).

The fm-api service is started by systemctl, while dnsmasq is managed by
SM. Even when dnsmasq is running, it may take some time before it can
answer DNS queries. To handle this, the fm-api now checks whether the
host FQDN can be resolved before starting the wsgi.Server.
With this fix, the workaround is not necessary anymore.

Tests performed:
- AIO-DX IPv4 fresh install
- AIO-DX IPv6 fresh install
- AIO-DX swact
- AIO-SX subcloud factory install

Depends-on: https://review.opendev.org/c/starlingx/fault/+/967753
Closes-Bug: 2116270

Change-Id: I1b93df45ed8cbfb4996c29e5fe622e6283299a43
Signed-off-by: Fabiano Correa Mercer <fabiano.correamercer@windriver.com>