Prevent cloud-init-seed service from starting too early
The cloud-init-seed.service was being started too early in the boot
process when Factory Install seed ISO is used. The Factory Install
process is completed successfully, but during the boot process, an error
message is logged in console log indicating that the cloud-init-seed
service failed to start.
The attempt to start that service should not occur at that time,
since that service should only start when the enrollment seed ISO is
detected.
In order to prevent that early start attempt, the cloud-init-seed
service unit file has been modified to include the condition that
checks for the presence of the Factory Install complete flag file.
That flag file is created at the end of the Factory Install process,
and its presence indicates that the Factory Install is completed.
It also helps us make decisions to execute other operations post-factory
install, such as enrollment.

Test Plan:
PASS - Prepare a Factory Install seed ISO and boot a subcloud with it.
     - Verify that the Factory Install process completes successfully
       and that there are no error messages in console log related to
       cloud-init-seed service.
     - Trigger enrollment process and verify that it works as expected.

Closes-Bug: 2132016

Change-Id: If977a8a873f1335e668b1623d39fa9c27e3e908c
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>