Match Baomon Config Options to CLI Options
Adding the ability to set timeout & log level values from CLI flags, and
adding the ability to set k8s & in-cluster values from config file. The
values from the flag takes precedence over the config file.

Test Plan:
PASS    Timeout option is set correctly when using the CLI flag
PASS    Log level option is set correctly when using the CLI flag
PASS    k8s option is set correctly when explicitly set in the config
file, and the flag was not raised
PASS    in-cluster option is set correctly when explicitly set in the
config file, and the flag was not raised

Story: 2011244
Task: 52705

Change-Id: I1d9f6f3009b88becdff240062b5f6221d74da704
Signed-off-by: Tae Park <tae.park@windriver.com>