collect: /var/crash not included in the bundle
If a kernel crash occurs files are written to the /var/crash directory.
These are text files that are are not large and a limited
number of files are kept as older ones are "rotated out" and deleted.

The useful information in those text files about kernel crashes is
considered to be worth the small amount of extra data in the bundle.

Files in /var/crash are now included in the collect bundle. If /var/crash
is empty no /var/crash directory will be bundled and
/var/extra/crash.info will report that the directory is empty.

Test Plan:

Preparation: on a system with at least 2 nodes generate a kernel crash
on each node.

   PASSED:- collect - single node collect.
   PASSED:- collect all - all nodes collect
   PASSED:- collect - empty /var/crash directory; no files copied

Closes-Bug: 2133228
Change-Id: Ic7b717ad58dda3b950d7f0e7ee3daf59f7cbad35
Signed-off-by: Tim Rose <kevin.rose@windriver.com>