Change to the ongoing execution validation
In order to avoid multiple instances of the tool, the script searches
for some known processes names in local and remote hosts. However, the
search for a very broad pattern was causing false positives, leading to
the tool to exit even when no other instance was running.

This change provides a more strict search pattern, focusing only on
the exact process names used by the main script (local and remote).

TEST PLAN:
PASS: Successfully execution of the collect command.
PASS: Successfully blocks new instances and notifies the user if collect
      is started while another already is running on the same node.
PASS: Successfully identifies and notifies the user if a new instance is
      started on some node (like stand-by controller or compute)
      while another instance is running in active controller (with --all
      option).
PASS: Successfully avoids false positives when other processes are
      running with similar names.

Closes-Bug: #2133935

Change-Id: Ia9a6e9a6683478ce149069fbd2d262171aef5257
Signed-off-by: Eduardo Alberti <eduardo.alberti@windriver.com>